POB_quiz_backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
PyJWT==2.9.0
web3==6.20.1
python-dotenv==1.0.1
django-cors-headers==4.4.0
psycopg[binary]==3.2.11
redis==5.0.8
pytest==8.3.3
pytest-django==4.9.0
freezegun==1.5.1
